Added prep work for branches, loading textures and setting positions and sides...Not animated yet
Wrote function prototype for a function called updateBranches. Declared an int constant called NUM_BRANCHES and initialized it to 6. Declared an array of Sprite objects called branches that can hold six
Sprite instances. Declared a enumberation called side with 3 possible vlaues: LEFT, RIGHT, and NONE. Prepared the branches and set the textures. Set the origin  and updated the branche sprites each frame. Drew the branches in the seen however not animated yet